runpod_v3/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
PyTurboJPEG>=1.7.1
